requirements.txt
ccxt
pandas
pandas-ta
python-dotenv
fastapi
uvicorn
tensorflow
scikit-learn
joblib
